Add build tag "integration" to integration test
The integration test runs containers and can be non-trivial to execute
