+++ internal/logservice/logserver_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package logservice
 
 import (
@@ -12,7 +14,6 @@ import (
 
 	"github.com/gocql/gocql"
 	"github.com/google/uuid"
-	"github.com/nlnwa/veidemann-api/go/commons/v1"
 	logV1 "github.com/nlnwa/veidemann-api/go/log/v1"
 	"github.com/nlnwa/veidemann-log-service/internal/scylla"
 	"github.com/nlnwa/veidemann-log-service/pkg/logservice"
@@ -23,7 +24,6 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/protobuf/encoding/protojson"
-	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 var (
@@ -57,9 +57,6 @@ func TestMain(m *testing.M) {
 				networkName: {"scylla"},
 			},
 			SkipReaper: true,
-			//HostConfigModifier: func(config *container.HostConfig) {
-			//	config.AutoRemove = true
-			//},
 			WaitingFor: wait.ForListeningPort("9042/tcp"),
 		},
 		Started: true,
@@ -77,9 +74,6 @@ func TestMain(m *testing.M) {
 				"CQLSH_HOST": "scylla",
 				"KEYSPACE":   keyspace,
 			},
-			//HostConfigModifier: func(config *container.HostConfig) {
-			//	config.AutoRemove = true
-			//},
 			WaitingFor: wait.ForLog("Schema initialized"),
 		},
 		Started: true,
@@ -373,103 +367,3 @@ func truncate(table string) {
 		time.Sleep(time.Millisecond)
 	}
 }
-
-func timestamp() *timestamppb.Timestamp {
-	return timestamppb.New(time.Now().UTC().Truncate(time.Millisecond))
-}
-
-var (
-	crawlLog1 = &logV1.CrawlLog{
-		ExecutionId:         uuid.NewString(),
-		JobExecutionId:      uuid.NewString(),
-		WarcId:              uuid.NewString(),
-		FetchTimeStamp:      timestamp(),
-		BlockDigest:         "sha1:f054ed8f9fd5893d6b70dc336a68e8092782723c",
-		CollectionFinalName: "Collection_2021",
-		ContentType:         "text/dns",
-		DiscoveryPath:       "P",
-		FetchTimeMs:         46,
-		IpAddress:           "8.8.8.8:53",
-		RecordType:          "response",
-		RequestedUri:        "dns:www.example.com",
-		Size:                50,
-		Error: &commons.Error{
-			Code:   -1,
-			Msg:    "Error",
-			Detail: "Everything went wrong",
-		},
-		StatusCode: 1,
-		StorageRef: "warcfile:Collection_2021-20210415110455-veidemann_contentwriter_775ffd88bc_5ljbb-00000.warc.gz:667",
-	}
-	crawlLog2 = &logV1.CrawlLog{
-		ExecutionId:         crawlLog1.ExecutionId,
-		JobExecutionId:      crawlLog1.JobExecutionId,
-		WarcId:              uuid.NewString(),
-		FetchTimeStamp:      timestamp(),
-		BlockDigest:         "sha1:f054ed8f9fd5893d6b70dc336a68e8092782723c",
-		CollectionFinalName: "Collection_2021",
-		ContentType:         "text/dns",
-		DiscoveryPath:       "P",
-		FetchTimeMs:         46,
-		IpAddress:           "8.8.8.8:53",
-		RecordType:          "response",
-		RequestedUri:        "dns:www.example.com",
-		Size:                50,
-		StatusCode:          200,
-		StorageRef:          "warcfile:Collection_2021-20210415110455-veidemann_contentwriter_775ffd88bc_5ljbb-00000.warc.gz:667",
-	}
-	pageLog1 = &logV1.PageLog{
-		ExecutionId:         uuid.NewString(),
-		WarcId:              uuid.NewString(),
-		JobExecutionId:      uuid.NewString(),
-		Uri:                 "https://www.nb.no/samlinger",
-		Referrer:            "https://www.nb.no/",
-		CollectionFinalName: "Veidemann_2021",
-		Method:              "GET",
-		Resource: []*logV1.PageLog_Resource{
-			{
-				Uri:           "https://www.nb.no/samlinger",
-				FromCache:     false,
-				Renderable:    false,
-				ResourceType:  "t",
-				ContentType:   "text/html",
-				StatusCode:    200,
-				DiscoveryPath: "L",
-				WarcId:        uuid.NewString(),
-				Referrer:      "https://www.nb.no/",
-				Error:         nil,
-				Method:        "GET",
-			},
-		},
-		Outlink: []string{
-			"https://www.nb.no/whatever",
-		},
-	}
-	pageLog2 = &logV1.PageLog{
-		WarcId:              uuid.NewString(),
-		ExecutionId:         pageLog1.ExecutionId,
-		JobExecutionId:      pageLog1.JobExecutionId,
-		Uri:                 "https://www.nb.no/presse",
-		Referrer:            "https://www.nb.no/",
-		CollectionFinalName: "Veidemann_2021",
-		Method:              "GET",
-		Resource: []*logV1.PageLog_Resource{
-			{
-				Uri:           "https://www.nb.no/presse",
-				FromCache:     false,
-				Renderable:    false,
-				ResourceType:  "t",
-				ContentType:   "text/html",
-				StatusCode:    200,
-				DiscoveryPath: "L",
-				WarcId:        uuid.NewString(),
-				Referrer:      "https://www.nb.no/",
-				Error:         nil,
-				Method:        "GET",
-			},
-		},
-		Outlink: []string{
-			"https://www.nb.no/foobar",
-		},
-	}
-)
