+++ internal/logservice/metrics_test.go
@@ -16,7 +16,15 @@
 
 package logservice
 
-import "testing"
+import (
+	"testing"
+	"time"
+
+	"github.com/google/uuid"
+	"github.com/nlnwa/veidemann-api/go/commons/v1"
+	logV1 "github.com/nlnwa/veidemann-api/go/log/v1"
+	"google.golang.org/protobuf/types/known/timestamppb"
+)
 
 func BenchmarkCollectCrawlLog(b *testing.B) {
 	for n := 0; n < b.N; n++ {
@@ -29,3 +37,103 @@ func BenchmarkCollectPageLog(b *testing.B) {
 		CollectPageLog(pageLog1)
 	}
 }
+
+func timestamp() *timestamppb.Timestamp {
+	return timestamppb.New(time.Now().UTC().Truncate(time.Millisecond))
+}
+
+var (
+	crawlLog1 = &logV1.CrawlLog{
+		ExecutionId:         uuid.NewString(),
+		JobExecutionId:      uuid.NewString(),
+		WarcId:              uuid.NewString(),
+		FetchTimeStamp:      timestamp(),
+		BlockDigest:         "sha1:f054ed8f9fd5893d6b70dc336a68e8092782723c",
+		CollectionFinalName: "Collection_2021",
+		ContentType:         "text/dns",
+		DiscoveryPath:       "P",
+		FetchTimeMs:         46,
+		IpAddress:           "8.8.8.8:53",
+		RecordType:          "response",
+		RequestedUri:        "dns:www.example.com",
+		Size:                50,
+		Error: &commons.Error{
+			Code:   -1,
+			Msg:    "Error",
+			Detail: "Everything went wrong",
+		},
+		StatusCode: 1,
+		StorageRef: "warcfile:Collection_2021-20210415110455-veidemann_contentwriter_775ffd88bc_5ljbb-00000.warc.gz:667",
+	}
+	crawlLog2 = &logV1.CrawlLog{
+		ExecutionId:         crawlLog1.ExecutionId,
+		JobExecutionId:      crawlLog1.JobExecutionId,
+		WarcId:              uuid.NewString(),
+		FetchTimeStamp:      timestamp(),
+		BlockDigest:         "sha1:f054ed8f9fd5893d6b70dc336a68e8092782723c",
+		CollectionFinalName: "Collection_2021",
+		ContentType:         "text/dns",
+		DiscoveryPath:       "P",
+		FetchTimeMs:         46,
+		IpAddress:           "8.8.8.8:53",
+		RecordType:          "response",
+		RequestedUri:        "dns:www.example.com",
+		Size:                50,
+		StatusCode:          200,
+		StorageRef:          "warcfile:Collection_2021-20210415110455-veidemann_contentwriter_775ffd88bc_5ljbb-00000.warc.gz:667",
+	}
+	pageLog1 = &logV1.PageLog{
+		ExecutionId:         uuid.NewString(),
+		WarcId:              uuid.NewString(),
+		JobExecutionId:      uuid.NewString(),
+		Uri:                 "https://www.nb.no/samlinger",
+		Referrer:            "https://www.nb.no/",
+		CollectionFinalName: "Veidemann_2021",
+		Method:              "GET",
+		Resource: []*logV1.PageLog_Resource{
+			{
+				Uri:           "https://www.nb.no/samlinger",
+				FromCache:     false,
+				Renderable:    false,
+				ResourceType:  "t",
+				ContentType:   "text/html",
+				StatusCode:    200,
+				DiscoveryPath: "L",
+				WarcId:        uuid.NewString(),
+				Referrer:      "https://www.nb.no/",
+				Error:         nil,
+				Method:        "GET",
+			},
+		},
+		Outlink: []string{
+			"https://www.nb.no/whatever",
+		},
+	}
+	pageLog2 = &logV1.PageLog{
+		WarcId:              uuid.NewString(),
+		ExecutionId:         pageLog1.ExecutionId,
+		JobExecutionId:      pageLog1.JobExecutionId,
+		Uri:                 "https://www.nb.no/presse",
+		Referrer:            "https://www.nb.no/",
+		CollectionFinalName: "Veidemann_2021",
+		Method:              "GET",
+		Resource: []*logV1.PageLog_Resource{
+			{
+				Uri:           "https://www.nb.no/presse",
+				FromCache:     false,
+				Renderable:    false,
+				ResourceType:  "t",
+				ContentType:   "text/html",
+				StatusCode:    200,
+				DiscoveryPath: "L",
+				WarcId:        uuid.NewString(),
+				Referrer:      "https://www.nb.no/",
+				Error:         nil,
+				Method:        "GET",
+			},
+		},
+		Outlink: []string{
+			"https://www.nb.no/foobar",
+		},
+	}
+)
